refactor: combine ContainerStatus definition in driver.
