+++ pkg/types/container.go
@@ -1,0 +1,13 @@
+package types
+
+type ContainerStatus string
+
+const (
+	StatusCreated    ContainerStatus = "created"
+	StatusRunning    ContainerStatus = "running"
+	StatusPaused     ContainerStatus = "paused"
+	StatusRestarting ContainerStatus = "restarting"
+	StatusRemoving   ContainerStatus = "removing"
+	StatusExited     ContainerStatus = "exited"
+	StatusDead       ContainerStatus = "dead"
+)
